+++ bin/update-deps.sh
@@ -5,3 +5,8 @@ uv sync --no-install-project --all-extras --upgrade
 uv tree --outdated
 npm update
 npm outdated
+if [ -d frontend ]; then
+  cd frontend
+  npm update
+  npm outdated
+fi