+++ picopt/report.py
@@ -6,7 +6,6 @@ from subprocess import CalledProcessError
 
 from confuse import AttrDict
 from humanize import naturalsize
-from termcolor import cprint
 
 from picopt.path import PathInfo
 from picopt.printer import Printer
@@ -30,10 +29,11 @@ class ReportStats(ReportStatBase):
 
     def __init__(
         self,
-        path,
+        path: Path,
         *args,
         config: AttrDict | None = None,
         path_info: PathInfo | None = None,
+        converted: bool = False,
         **kwargs,
     ) -> None:
         """Initialize required instance variables."""
@@ -44,6 +44,7 @@ class ReportStats(ReportStatBase):
         self._full_name = path_info.full_output_name() if path_info else str(path)
         super().__init__(path, *args, **kwargs)
         self.saved = self.bytes_in - self.bytes_out
+        self.converted = converted
 
     def _new_percent_saved(self) -> str:
         """Spit out how much space the optimization saved."""
@@ -86,26 +87,31 @@ class ReportStats(ReportStatBase):
 
     def report(self, printer: Printer) -> None:
         """Record the percent saved & print it."""
+        # Pass the printer in at the end here to avoid pickling
         if self.exc:
             report = self._report_error()
             printer.error(report, self.exc)
             return
         report = self._report_saved()
         if self.saved > 0:
-            printer.saved_message(report)
+            if self.converted:
+                printer.converted(report)
+            else:
+                printer.saved(report)
         else:
-            printer.lost_message(report)
+            printer.lost(report)
 
 
 class Totals:
     """Totals for final report."""
 
-    def __init__(self, config: AttrDict):
+    def __init__(self, config: AttrDict, printer: Printer):
         """Initialize Totals."""
         self.bytes_in: int = 0
         self.bytes_out: int = 0
         self.errors: list[ReportStats] = []
         self._config: AttrDict = config
+        self._printer: Printer = printer
 
     def _report_bytes_in(self) -> None:
         """Report Totals if there were bytes in."""
@@ -129,20 +135,20 @@ class Totals:
             msg = "Lost"
         natural_saved = naturalsize(bytes_saved)
         msg += f" a total of {natural_saved} or {percent_bytes_saved:.2f}%"
-        cprint(msg)
+        self._printer.saved(msg)
         if self._config.dry_run:
-            cprint("Dry run did not change any files.")
+            self._printer.final_message("Dry run did not change any files.")
 
-    def report(self, printer: Printer) -> None:
+    def report(self) -> None:
         """Report the total number and percent of bytes saved."""
         if self._config.verbose == 1:
-            cprint("")
+            print("")  # noqa: T201
         if self.bytes_in:
             self._report_bytes_in()
-        elif self._config.verbose:
-            cprint("Didn't optimize any files.")
+        else:
+            self._printer.final_message("Didn't optimize any files.")
 
         if self.errors:
-            cprint("Errors with the following files:", "red")
+            self._printer.error_title("Errors with the following files:")
             for rs in self.errors:
-                rs.report(printer)
+                rs.report(self._printer)