+++ tests/integration/test_images_dir.py
@@ -54,7 +54,7 @@ if platform.system() == "Darwin":
 else:
     FNS.update(
         {
-            "test_animated_png.png": (63435, 63058, ("png", 63058), ("webp", 53422)),
+            "test_animated_png.png": (63435, 63058, ("png", 63058), ("webp", 52864)),
             "test_gif.gif": (138952, 138944, ("png", 112290), ("webp", 107924)),
             "test_pre-optimized_png.png": (
                 256572,